Also run the vitest tests.
