+++ src/package.json
@@ -68,13 +68,13 @@
     "socket.io": "^4.7.5",
     "socket.io-client": "^4.7.5",
     "superagent": "10.0.2",
-    "threads": "^1.7.0",
     "tinycon": "0.6.8",
     "tsx": "4.17.0",
     "ueberdb2": "^4.2.93",
     "underscore": "1.13.7",
     "unorm": "1.6.0",
-    "wtfnode": "^0.9.3"
+    "wtfnode": "^0.9.3",
+    "vitest": "^2.0.5"
   },
   "bin": {
     "etherpad-healthcheck": "../bin/etherpad-healthcheck",
@@ -89,6 +89,7 @@
     "@types/jquery": "^3.5.30",
     "@types/jsdom": "^21.1.7",
     "@types/jsonwebtoken": "^9.0.6",
+    "@types/mime-types": "^2.1.4",
     "@types/mocha": "^10.0.7",
     "@types/node": "^22.3.0",
     "@types/oidc-provider": "^8.5.2",
@@ -132,7 +133,8 @@
     "test-ui:ui": "cross-env NODE_ENV=production npx playwright test tests/frontend-new/specs --ui",
     "test-admin": "cross-env NODE_ENV=production npx playwright test tests/frontend-new/admin-spec --workers 1",
     "test-admin:ui": "cross-env NODE_ENV=production npx playwright test tests/frontend-new/admin-spec --ui --workers 1",
-    "debug:socketio": "cross-env DEBUG=socket.io* node --require tsx/cjs node/server.ts"
+    "debug:socketio": "cross-env DEBUG=socket.io* node --require tsx/cjs node/server.ts",
+    "test:vitest": "vitest"
   },
   "version": "2.2.2",
   "license": "Apache-2.0"
